changes based on igels review
