+++ templates/layout.html
@@ -38,7 +38,7 @@
                 · <a href="https://github.com/engelsystem/engelsystem/issues">Bugs / Features</a>
                 · <a href="https://github.com/engelsystem/engelsystem/">Development Platform</a>
                 · <a href="%credits_url%">Credits</a>
-            </div>;
+            </div>
         </div>
     </div>
 </div>
